Phase II VRS instructions
Fix. 
On this page, under item 3 – there are two items that need to have their own bullets under the VRS bullet.  It is confusing because these are not sub-bullets to VRS but standalone statements.
https://seedfund.nsf.gov/resources/awardees/phase-2/instructions/

•	VRS bottom line for the overall budget is either $10,000 for one, $20,000 for two, or $30,000 for three. (Funds should only appear on budget line F.1)
- CAP Budget Info: All costs allocated to the service provider are entered under line G.6 as “other costs”.
- For all other supplements: please refer to the DCL for budgetary guidance.

And the header for #3 should be modified to include “CAP” like  RAHSS/REU/RET/VRS/CAP Budget Info