Clarified Fastlane v Research.gov
Hello,

I just wanted to bring to your attention that we have started to receive some emails from submitters that are trying to submit to the SBIR program via Research.Gov.  

Although the FastLane guide you provided to them on the Seedfund site states to go to FastLane, there is still some confusion and some submitters are going to Research.Gov.  I was thinking maybe if adding something to the Advisory on Fastlane or the Seedfund site splash page advising applicants to only use Fastlane would steer them in the right direction.

I think closer and closer we get to the deadline we are going to get more emails and calls from applicants trying to use Research.Gov for submission.

Either way I just wanted you to know what type of calls we have been seeing.

Thanks,

Jorge Urrutia – Technical Support  (contractor)
IT Help Central Services